Change profile edit button to be common
