+++ src/components/common/EditButton/style.module.scss.d.ts
