+++ src/components/common/index.ts
@@ -4,6 +4,7 @@ export { default as CommunityLogo } from './CommunityLogo';
 export { default as Cropper } from './Cropper';
 export { DRAG_HANDLE, default as Draggable } from './Draggable';
 export { default as Dropdown } from './Dropdown';
+export { default as EditButton } from './EditButton';
 export { default as GifSafeImage } from './GifSafeImage';
 export { default as LinkButton } from './LinkButton';
 export { default as Modal } from './Modal';
