+++ src/components/store/ItemHeader/index.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@/components/common';
 import Diamonds from '@/components/store/Diamonds';
-import EditButton from '@/components/store/EditButton';
 import HiddenIcon from '@/components/store/HiddenIcon';
+import StoreEditButton from '@/components/store/StoreEditButton';
 import styles from './style.module.scss';
 
 interface ItemHeaderProps {
@@ -25,7 +25,7 @@ const ItemHeader = ({
     <div className={styles.itemHeaderGroup}>
       <Typography variant="h1/bold" className={styles.itemName}>
         <span>{itemName}</span>
-        {uuid && showEdit ? <EditButton type="item" uuid={uuid} /> : null}
+        {uuid && showEdit ? <StoreEditButton type="item" uuid={uuid} /> : null}
         {isHidden ? <HiddenIcon type="item" /> : null}
       </Typography>
       {cost === undefined ? null : (
