+++ src/components/store/index.ts
@@ -5,7 +5,6 @@ export { default as CartItemCard } from './CartItemCard';
 export { default as CollectionSlider } from './CollectionSlider';
 export { default as CreateButton } from './CreateButton';
 export { default as Diamonds } from './Diamonds';
-export { default as EditButton } from './EditButton';
 export { default as HelpModal } from './HelpModal';
 export { default as Hero } from './Hero';
 export { default as HiddenIcon } from './HiddenIcon';
@@ -15,4 +14,5 @@ export { default as OrdersDisplay } from './OrdersDisplay';
 export { default as PickupEventDetail } from './PickupEventDetail';
 export { default as PickupEventPicker } from './PickupEventPicker';
 export { default as StoreConfirmModal } from './StoreConfirmModal';
+export { default as StoreEditButton } from './StoreEditButton';
 export { default as Navbar } from './StoreNavbar';
